Sagar_Backend/time based strategy/strategy_optimization/requirements.txt
fastapi==0.115.2
mysql_connector_repackaged==0.3.1
pydantic==2.9.2
Requests==2.32.3
uvicorn==0.31.1
